feat(#394): refactor code color
