+++ studio/src/app/popovers/editor/app-color/app-color.scss
@@ -1,5 +1,6 @@
 app-color {
 
+  @import "../../../../global/theme/editor/editor-info";
   @import "../../../../global/theme/editor/editor-popover";
 
   deckgo-color {
