+++ studio/src/app/popovers/editor/app-color/app-color.tsx
@@ -3,6 +3,7 @@ import {Component, Element, Event, EventEmitter, h, Prop, State} from '@stencil/
 import {isIPad} from '@deckdeckgo/utils';
 
 import {TargetElement} from '../../../utils/editor/target-element';
+import {SlotType} from '../../../utils/editor/slot-type';
 
 @Component({
     tag: 'app-color',
@@ -35,14 +36,19 @@ export class AppColor {
     @State()
     private poll: boolean = false;
 
+    @State()
+    private code: boolean = false;
+
     async componentWillLoad() {
         if (this.deckOrSlide) {
             this.qrCode = this.selectedElement && this.selectedElement.tagName && this.selectedElement.tagName.toUpperCase() === 'deckgo-slide-qrcode'.toUpperCase();
             this.chart = this.selectedElement && this.selectedElement.tagName && this.selectedElement.tagName.toUpperCase() === 'deckgo-slide-chart'.toUpperCase();
             this.poll = this.selectedElement && this.selectedElement.tagName && this.selectedElement.tagName.toUpperCase() === 'deckgo-slide-poll'.toUpperCase();
-
-            this.applyToTargetElement = this.qrCode || this.poll ? TargetElement.QR_CODE : (this.chart ? TargetElement.CHART : TargetElement.SLIDE);
         }
+
+        this.code = this.selectedElement && this.selectedElement.nodeName && this.selectedElement.nodeName.toLocaleLowerCase() === SlotType.CODE;
+
+        this.applyToTargetElement = this.code ? TargetElement.CODE : (this.qrCode || this.poll ? TargetElement.QR_CODE : (this.chart ? TargetElement.CHART : TargetElement.SLIDE));
 
         this.moreColors = !isIPad();
     }
@@ -86,7 +92,7 @@ export class AppColor {
                 <ion-icon name="close"></ion-icon>
             </ion-router-link>
         </ion-toolbar>,
-            <app-select-target-element deckOrSlide={this.deckOrSlide} qrCode={this.qrCode || this.poll} chart={this.chart || this.poll}
+            <app-select-target-element deckOrSlide={this.deckOrSlide} qrCode={this.qrCode || this.poll} chart={this.chart || this.poll} code={this.code}
                                        onApplyTo={($event: CustomEvent<TargetElement>) => this.selectApplyToTargetElement($event)}></app-select-target-element>,
 
             this.renderColorOptions()
@@ -102,11 +108,15 @@ export class AppColor {
             return <app-color-chart selectedElement={this.selectedElement}
                                      onColorChange={($event: CustomEvent<boolean>) => this.colorChange($event)}
                                      moreColors={this.moreColors}></app-color-chart>
+        } else if (this.applyToTargetElement === TargetElement.CODE) {
+            return <app-color-code selectedElement={this.selectedElement}
+                                   onColorChange={($event: CustomEvent<boolean>) => this.colorChange($event)}
+                                   moreColors={this.moreColors}></app-color-code>
         } else {
             return <app-color-deck-slide selectedElement={this.selectedElement} moreColors={this.moreColors}
-                                            deckOrSlide={this.deckOrSlide}
-                                            onColorChange={($event: CustomEvent<boolean>) => this.colorChange($event)}
-                                            applyToAllDeck={this.applyToTargetElement === TargetElement.DECK}></app-color-deck-slide>
+                                         deckOrSlide={this.deckOrSlide}
+                                         onColorChange={($event: CustomEvent<boolean>) => this.colorChange($event)}
+                                         applyToAllDeck={this.applyToTargetElement === TargetElement.DECK}></app-color-deck-slide>
         }
     }
 }
