+++ studio/src/app/utils/editor/target-element.tsx
@@ -2,5 +2,7 @@ export enum TargetElement {
     QR_CODE= 'qr_code',
     CHART = 'chart',
     SLIDE = 'slide',
-    DECK = 'deck'
+    DECK = 'deck',
+    SECTION = 'section',
+    CODE = 'code'
 }
