+++ studio/src/components.d.ts
@@ -50,6 +50,10 @@ export namespace Components {
     'moreColors': boolean;
     'selectedElement': HTMLElement;
   }
+  interface AppColorCode {
+    'moreColors': boolean;
+    'selectedElement': HTMLElement;
+  }
   interface AppColorDeckSlide {
     'applyToAllDeck': boolean;
     'deckOrSlide': boolean;
@@ -181,6 +185,7 @@ export namespace Components {
   interface AppRoot {}
   interface AppSelectTargetElement {
     'chart': boolean;
+    'code': boolean;
     'deckOrSlide': boolean;
     'qrCode': boolean;
   }
@@ -250,6 +255,12 @@ declare global {
   var HTMLAppColorChartElement: {
     prototype: HTMLAppColorChartElement;
     new (): HTMLAppColorChartElement;
+  };
+
+  interface HTMLAppColorCodeElement extends Components.AppColorCode, HTMLStencilElement {}
+  var HTMLAppColorCodeElement: {
+    prototype: HTMLAppColorCodeElement;
+    new (): HTMLAppColorCodeElement;
   };
 
   interface HTMLAppColorDeckSlideElement extends Components.AppColorDeckSlide, HTMLStencilElement {}
@@ -660,6 +671,7 @@ declare global {
     'app-code': HTMLAppCodeElement;
     'app-color': HTMLAppColorElement;
     'app-color-chart': HTMLAppColorChartElement;
+    'app-color-code': HTMLAppColorCodeElement;
     'app-color-deck-slide': HTMLAppColorDeckSlideElement;
     'app-color-qrcode': HTMLAppColorQrcodeElement;
     'app-contact': HTMLAppContactElement;
@@ -749,6 +761,11 @@ declare namespace LocalJSX {
     'selectedElement'?: HTMLElement;
   }
   interface AppColorChart {
+    'moreColors'?: boolean;
+    'onColorChange'?: (event: CustomEvent<boolean>) => void;
+    'selectedElement'?: HTMLElement;
+  }
+  interface AppColorCode {
     'moreColors'?: boolean;
     'onColorChange'?: (event: CustomEvent<boolean>) => void;
     'selectedElement'?: HTMLElement;
@@ -908,6 +925,7 @@ declare namespace LocalJSX {
   interface AppRoot {}
   interface AppSelectTargetElement {
     'chart'?: boolean;
+    'code'?: boolean;
     'deckOrSlide'?: boolean;
     'onApplyTo'?: (event: CustomEvent<TargetElement>) => void;
     'qrCode'?: boolean;
@@ -952,6 +970,7 @@ declare namespace LocalJSX {
     'app-code': AppCode;
     'app-color': AppColor;
     'app-color-chart': AppColorChart;
+    'app-color-code': AppColorCode;
     'app-color-deck-slide': AppColorDeckSlide;
     'app-color-qrcode': AppColorQrcode;
     'app-contact': AppContact;
@@ -1034,6 +1053,7 @@ declare module "@stencil/core" {
       'app-code': LocalJSX.AppCode & JSXBase.HTMLAttributes<HTMLAppCodeElement>;
       'app-color': LocalJSX.AppColor & JSXBase.HTMLAttributes<HTMLAppColorElement>;
       'app-color-chart': LocalJSX.AppColorChart & JSXBase.HTMLAttributes<HTMLAppColorChartElement>;
+      'app-color-code': LocalJSX.AppColorCode & JSXBase.HTMLAttributes<HTMLAppColorCodeElement>;
       'app-color-deck-slide': LocalJSX.AppColorDeckSlide & JSXBase.HTMLAttributes<HTMLAppColorDeckSlideElement>;
       'app-color-qrcode': LocalJSX.AppColorQrcode & JSXBase.HTMLAttributes<HTMLAppColorQrcodeElement>;
       'app-contact': LocalJSX.AppContact & JSXBase.HTMLAttributes<HTMLAppContactElement>;
